Move interactive-tx funding to a dedicated actor
This commit refactors the funding steps of an interactive-tx to a separate
actor: it doesn't contain any business logic changes.
